requirements.txt
langchain_community==0.3.18
langchain_core==0.3.40
langchain_groq==0.2.4
python-dotenv==1.0.1
streamlit==1.42.1
